Opzione D: refactor URL probe/scaffold
- cmd_url_inspect.py → _url_probe.py (rename, solo HTTP probe)
- generate_yaml_scaffold() spostato in _url_scout_common.py (puro)
- _scout() riusa probe_url() per validazione (HEAD+fallback, -15 righe)
- _generate_dataset_yml() usa _generate_raw_sources_block() condiviso
- -206 righe nette, test 662/662
