+++ toolkit/cli/_url_scout_common.py
@@ -1,7 +1,6 @@
-"""Helper condivisi per URL scout tra init --url e inspect url.
+"""Helper condivisi per URL scout — slug, estensione, filename, scaffold YAML.
 
-Consolida slug generation, inferenza estensione e filename
-per evitare divergenze tra cmd_init.py e cmd_url_inspect.py.
+Tutta logica pura (no HTTP, no I/O). Condivisa tra init --url e inspect url.
 """
 
 from __future__ import annotations
@@ -9,6 +8,7 @@ from __future__ import annotations
 import re
 import uuid
 from pathlib import Path
+from typing import Any
 from urllib.parse import urlparse
 
 
@@ -52,3 +52,110 @@ def infer_filename(url: str, slug: str) -> str:
         return Path(path).stem + ".csv"
     name = Path(path).name
     return name or f"{slug}.csv"
+
+
+# ---------------------------------------------------------------------------
+# Scaffold YAML — raw.sources block
+# ---------------------------------------------------------------------------
+
+
+def _make_source_name(link: str) -> str:
+    stem = Path(urlparse(link).path).stem
+    return re.sub(r"[^a-z0-9_]", "_", (stem or "resource").lower())
+
+
+def _infer_type_from_url(u: str) -> str:
+    if "/datastore/dump/" in u or "/datastore_search" in u:
+        return "ckan"
+    if "sdmx" in u.lower() or "/dataflow/" in u.lower():
+        return "sdmx"
+    return "http_file"
+
+
+def _generate_raw_sources_block(url: str, slug: str, fname: str | None = None) -> list[str]:
+    """Genera il blocco YAML raw.sources per un singolo URL.
+
+    Args:
+        url: URL della fonte dati.
+        slug: Slug del dataset (usato per source name).
+        fname: Nome file esplicito. Se None, inferito dall'URL.
+    """
+    parsed = urlparse(url)
+    if fname is None:
+        fname = Path(parsed.path).name or f"{slug}.csv"
+    stype = _infer_type_from_url(url)
+    lines = [
+        f'    - name: "{slug}_source"',
+        f'      type: "{stype}"',
+        "      args:",
+        f'        url: "{url}"',
+        f'        filename: "{fname}"',
+    ]
+    lines.append("      primary: true")
+    return lines
+
+
+def generate_yaml_scaffold(
+    probe_result: dict[str, Any],
+    ckan_resources: list[dict[str, Any]] | None = None,
+    candidate_links: list[str] | None = None,
+) -> str:
+    """Genera scaffold YAML per un URL ispezionato.
+
+    Usato da inspect url --scaffold. Il probe_result deve contenere almeno
+    final_url. Opzionalmente si possono passare risorse CKAN o link candidati
+    per popolare raw.sources in modo strutturato.
+    """
+    url = probe_result["final_url"]
+    parsed = urlparse(url)
+    slug = slugify(url)
+    lines = [
+        "# Scaffold generato da scout-url --scaffold",
+        "# Verifica e modifica prima di usare",
+        "",
+        'root: "../../out"',
+        "schema_version: 1",
+        "",
+        "dataset:",
+        f'  name: "{slug}"',
+        "  years: [2024]  # TBD: inferito da URL o sorgente",
+        "",
+        "raw:",
+        "  output_policy: overwrite",
+        "  sources:",
+    ]
+
+    if ckan_resources:
+        for res in ckan_resources:
+            res_name = re.sub(r"[^a-z0-9_]", "_", (res["name"] or "resource").lower())
+            res_url = res["url"]
+            fmt = res["format"]
+            fname = Path(urlparse(res_url).path).name or f"{res_name}.{fmt}"
+            portal_base = f"{parsed.scheme}://{parsed.netloc}"
+            lines.append(f'    - name: "{res_name}"')
+            lines.append('      type: "ckan"')
+            lines.append("      args:")
+            lines.append(f'        portal_url: "{portal_base}"')
+            lines.append(f'        resource_id: "{res.get("id") or ""}"')
+            lines.append(f'        filename: "{fname}"')
+            lines.append("      primary: true")
+    elif candidate_links:
+        seen: set[str] = set()
+        for link in candidate_links:
+            if link in seen:
+                continue
+            seen.add(link)
+            link_name = _make_source_name(link)
+            fname = Path(urlparse(link).path).name
+            stype = _infer_type_from_url(link)
+            lines.append(f'    - name: "{link_name}"')
+            lines.append(f'      type: "{stype}"')
+            lines.append("      args:")
+            lines.append(f'        url: "{link}"')
+            if fname:
+                lines.append(f'        filename: "{fname}"')
+            lines.append("      primary: true")
+    else:
+        lines.extend(_generate_raw_sources_block(url, slug))
+    lines.append("")
+    return "\n".join(lines)
